+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # Changelog for prairie
 
+## 0.1.0.0
+
+- [#20](https://github.com/parsonsmatt/prairie/pull/20)
+    - Add `tabulateRecordApply` which is like `tabulateRecordA` but allows you to use non-Applicaive effects.
+    - `Prairie.Fold`: `foldMapRecord` now works with any `Semigroup` instead of any `Monoid`, allowing you to fold into a `NonEmpty` list.
+    - `Prairie.Traverse`:
+        - `traverseRecord_` introduced as a slightly more efficient form of `void . traverseRecord f`.
+        - `traverseFields` and `traverseFields_` introduced to only operate on fields, not requiring a whole record.
+        - `traverseFieldsWithDict` (and `_` suffix variant) introduced to operate on fields with a constraint available.
+
 ## 0.0.4.1
 
 - [#18](https://github.com/parsonsmatt/prairie/pull/18)